requirements.txt
asgiref==3.5.2
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==5.3.0
